Avoid use of char as array index
ARMClang generates warning "array subscript is of type 'char'
[-Wchar-subscripts]", but also the Arm C library implements isprint() as
a macro which uses its argument as an array index. Since 'ch' would
normally be implicitly widened to an 'int' due to the prototype of
isprint(), as well as by being passed to variadic function tu_printf(),
we can avoid this by explicitly promoting its type to 'int'.